backend/requirements.txt
fastapi
uvicorn
scapy
python-dotenv
websockets
aiofiles